\name{C_dif}
\alias{C_dif}

\title{Native C Interface Wrapper for the Function \code{dif}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{dif} performs Deep Isolation Forest computations.
}

%\usage{
%.C(C_dif, res, dta, dimD, nt, nss)
%}
%
%\arguments{
%  \item{res}{An empty vector used to store the results.}
%  \item{dta}{A matrix of data (assumed to be stored in column‑major format).}
%  \item{dimD}{The number of rows and columns of \code{dta}.}
%  \item{nt}{The number of deep isolation trees to grow.}
%  \item{nss}{The number of subsamples used for training.}
%}

\keyword{internal}

