\name{C_getNCores}
\alias{C_getNCores}

\title{Native C Interface Wrapper for the Function \code{getNCores}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{getNCores} retrieves the maximum number of available CPU cores.
}

%\usage{
%.C(C_getNCores, n)
%}
%
%\arguments{
%  \item{n}{An integer that will receive the number of CPU cores.}
%}

\keyword{internal}

