\name{C_getNThreads}
\alias{C_getNThreads}

\title{Native C Interface Wrapper for the Function \code{getNThreads}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{getNThreads} retrieves the number of processing threads that should be used.
}

%\usage{
%.C(C_getNThreads, n)
%}
%
%\arguments{
%  \item{n}{An integer that will receive the number of threads.}
%}

\keyword{internal}

