\name{C_history_check}
\alias{C_history_check}

\title{Native C Interface Wrapper for the Function \code{history_check}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{history_check} scores non-positive values and historical outliers.
}

%\usage{
%.C(C_history_check, hScore, zScore, x, w, n)
%}
%
%\arguments{
%  \item{hScore}{An empty vector (for storing the output of historical outliers).}
%  \item{zScore}{An empty vector (for storing the output of zeros or negative values).}
%  \item{x}{A vector of current data.}
%  \item{w}{A vector of previous data.}
%  \item{n}{The length of the data vectors.}
%}

\keyword{internal}

