\name{C_history_res}
\alias{C_history_res}

\title{Native C Interface Wrapper for the Function \code{history_res}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{history_res} computes the residuals to identify historical outliers and to distinguish zero/negative values.
}

%\usage{
%.C(C_history_res, hRes, zScore, x, w, n)
%}
%
%\arguments{
%  \item{hRes}{An empty vector (for storing the output of historical residuals).}
%  \item{zScore}{An empty vector (for storing the output of zeros or negative values).}
%  \item{x}{A vector of current data.}
%  \item{w}{A vector of previous data.}
%  \item{n}{The length of the data vectors.}
%}

\keyword{internal}

