\name{C_openMP_version}
\alias{C_openMP_version}

\title{Native C Interface Wrapper for the Function \code{openMP_version}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{openMP_version} returns the OpenMP runtime version as a numeric value. It returns a numeric value containing the OpenMP version number (e.g., \code{2015.11}).
}

%\usage{
%.Call(C_openMP_version)
%}

\keyword{internal}

