\name{C_relat_check}
\alias{C_relat_check}

\title{Native C Interface Wrapper for the Function \code{relat_check}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{relat_check} checks for relational outliers.
}

%\usage{
%.C(C_relat_check, A, dim)
%}
%
%\arguments{
%  \item{A}{An input matrix.}
%  \item{dim}{The dimensions of the input matrix (i.e., a vector with the number of rows and columns of \code{A}).}
%}

\keyword{internal}

