\name{C_relat_res}
\alias{C_relat_res}

\title{Native C Interface Wrapper for the Function \code{relat_res}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{relat_res} computes the residuals from linear models
  to check relational outliers.
}

%\usage{
%.C(C_relat_res, A, dim)
%}
%
%\arguments{
%  \item{A}{An input matrix.}
%  \item{dim}{The dimensions of the input matrix (i.e., a vector with the number of rows and columns).}
%}
%
\keyword{internal}

