\name{C_setNThreads}
\alias{C_setNThreads}

\title{Native C Interface Wrapper for the Function \code{setNThreads}}

\description{
  These compiled symbols represent the direct entry points to the
  underlying C source code, masked behind the \code{C_} namespace fix.

  The function \code{setNThreads} sets the number of processing threads that should be used.
}

%\usage{
%.C(C_setNThreads, n)
%}
%
%\arguments{
%  \item{n}{An integer variable containing the desired number of threads.}
%}
%
\keyword{internal}

